--- lib/arroba/bsky/feed.rb
@@ -1,46 +1,0 @@
-# frozen_string_literal: true
-
-require_relative '../validations'
-
-module Arroba
-  class BSky
-    # The Feed resource provides access to posts, likes, reposts, and other
-    class Feed < BaseResource
-      include Validations::Limitable
-      # NOTE: it is described as not needing auth so with_auth is set to false, but it has given me 400 for that.
-      # When setting with_auth to true, it has only 404'd for me
-      get_with_query_params :describe_feed_generator, with_auth: false
-      get_with_enforced_limit :get_actor_feeds, :actor
-      get_with_enforced_limit :get_actor_likes, :actor
-      get_with_enforced_limit :get_author_feed, :actor, filter: nil, include_pins: nil, with_auth: false
-      get_with_query_params :get_feed_generator, :feed
-      get_with_query_params :get_feed_generators, :feeds
-      get_with_enforced_limit :get_feed_skeleton, :feed, with_auth: false
-      get_with_enforced_limit :get_feed, :feed
-      get_with_enforced_limit :get_likes, :uri, cid: nil
-      get_with_enforced_limit :get_list_feed, :list, with_auth: false
-      get_with_query_params :get_post_thread, :uri, depth: nil, parent_height: nil,
-                                                    with_auth: false do |depth: nil, parent_height: nil, **|
-        enforce_limit! depth, min: 0, max: 1000, name: 'Depth'
-        enforce_limit! parent_height, min: 0, max: 1000, name: 'Parent height'
-      end
-      get_with_query_params :get_posts, :uris do |uris:, **|
-        raise ArgumentError, 'URI list must be 25 or fewer items' if uris.count > 25
-      end
-      get_with_enforced_limit :get_quotes, :uri, cid: nil
-      get_with_enforced_limit :get_reposted_by, :uri, cid: nil
-      get_with_enforced_limit :get_suggested_feeds
-      get_with_enforced_limit :get_timeline, algorithm: nil
-      get_with_enforced_limit :search_posts, :q, sort: nil, since: nil, until: nil, mentions: nil, author: nil,
-                                                 lang: nil, domain: nil, url: nil,
-                                                 tag: nil do |sort: nil, tag: nil, **|
-        raise ArgumentError, 'sort must be "top" or "latest"' if !sort.nil? && sort != 'top' && sort != 'latest'
-        raise ArgumentError, 'tag elements must be shorter than 640 characters' if tag&.any? { |t| t.length > 640 }
-      end
-
-      def send_interactions!(interactions:)
-        post body: { interactions: }
-      end
-    end
-  end
-end
