--- lib/arroba/bsky/labeler.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module Arroba
-  class BSky
-    class Labeler < BaseResource
-      get_with_query_params :get_services, :dids, detailed: nil
-    end
-  end
-end
